requirements.txt
aiogram==3.25.0
python-dotenv==1.0.1
Pillow==11.1.0
img2pdf==0.6.0
redis==5.2.1
aiofiles==24.1.0
